[sergio.MeMenu] Adds "Me" storyboard to move the logout button into the Me modal menu. 1/2.
